FILE: organizes contents in each data-access lib
