+++ libs/menu-matriarch/dishes/data-access/src/lib/internal/dish-dto.ts
