+++ libs/menu-matriarch/ingredients/data-access/src/index.ts
@@ -1,5 +1,5 @@
-export * from './lib/ingredient.service';
-export * from './lib/ingredient-type.service';
 export * from './lib/create-ingredient-dto';
 export * from './lib/create-ingredient-type-dto';
+export * from './lib/ingredient.service';
+export * from './lib/ingredient-type.service';
 export * from './lib/mock-data';
