+++ libs/menu-matriarch/ingredients/data-access/src/lib/create-ingredient-dto.ts
@@ -1,4 +1,4 @@
-import { IngredientDto } from './ingredient-dto';
+import { IngredientDto } from './internal/ingredient-dto';
 
 export function createIngredientDto({
   id,
