+++ libs/menu-matriarch/ingredients/data-access/src/lib/create-ingredient-type-dto.ts
@@ -1,4 +1,4 @@
-import { IngredientTypeDto } from './ingredient-type-dto';
+import { IngredientTypeDto } from './internal/ingredient-type-dto';
 
 export function createIngredientTypeDto({
   id,
