+++ libs/menu-matriarch/ingredients/data-access/src/lib/internal/map-ingredient-type-dto-to-ingredient-type.spec.ts
@@ -1,5 +1,5 @@
 import { mapIngredientTypeDtoToIngredientType } from './map-ingredient-type-dto-to-ingredient-type';
-import { INGREDIENT_DTOS, INGREDIENT_TYPE_DTOS } from './mock-data';
+import { INGREDIENT_DTOS, INGREDIENT_TYPE_DTOS } from '../mock-data';
 
 describe('mapIngredientTypeDtoToIngredientType', () => {
   it('returns an ingredientType when passed ingredients', () => {
