+++ libs/menu-matriarch/meals/data-access/src/index.ts
@@ -1,3 +1,3 @@
-export * from './lib/meal.service';
 export * from './lib/create-meal-dto';
+export * from './lib/meal.service';
 export * from './lib/mock-data';
