+++ libs/menu-matriarch/meals/data-access/src/lib/create-meal-dto.ts
@@ -1,4 +1,4 @@
-import { MealDto } from './meal-dto';
+import { MealDto } from './internal/meal-dto';
 
 export function createMealDto({
   id,
