+++ libs/menu-matriarch/meals/data-access/src/lib/meal.service.ts
@@ -8,8 +8,11 @@ import { DishService } from '@atocha/menu-matriarch/dishes/data-access';
 import { RouterService } from '@atocha/menu-matriarch/shared/data-access-routing';
 import { Meal } from '@atocha/menu-matriarch/shared/util';
 import { TagService } from '@atocha/menu-matriarch/tags/data-access';
-import { EditableMealData, MealDataService } from './meal-data.service';
-import { mapMealDtoToMeal } from './map-meal-dto-to-meal';
+import {
+  EditableMealData,
+  MealDataService,
+} from './internal/meal-data.service';
+import { mapMealDtoToMeal } from './internal/map-meal-dto-to-meal';
 
 export type MealData = EditableMealData;
