+++ libs/menu-matriarch/menus/data-access/src/index.ts
@@ -1,4 +1,4 @@
+export * from './lib/create-menu-dto';
 export * from './lib/menu.service';
 export * from './lib/print.service';
-export * from './lib/create-menu-dto';
 export * from './lib/mock-data';
