+++ libs/menu-matriarch/menus/data-access/src/lib/create-menu-dto.ts
@@ -1,4 +1,4 @@
-import { MenuDto } from './menu-dto';
+import { MenuDto } from './internal/menu-dto';
 
 export function createMenuDto({
   id,
