+++ libs/menu-matriarch/menus/data-access/src/lib/internal/menu-dto.ts
