+++ libs/menu-matriarch/menus/data-access/src/lib/menu.service.ts
@@ -8,8 +8,11 @@ import { DishService } from '@atocha/menu-matriarch/dishes/data-access';
 import { UserService } from '@atocha/menu-matriarch/settings/data-access';
 import { RouterService } from '@atocha/menu-matriarch/shared/data-access-routing';
 import { Day, Menu } from '@atocha/menu-matriarch/shared/util';
-import { EditableMenuData, MenuDataService } from './menu-data.service';
-import { mapMenuDtoToMenu } from './map-menu-dto-to-menu';
+import {
+  EditableMenuData,
+  MenuDataService,
+} from './internal/menu-data.service';
+import { mapMenuDtoToMenu } from './internal/map-menu-dto-to-menu';
 
 @Injectable({
   providedIn: 'root',
