+++ libs/menu-matriarch/settings/data-access/src/index.ts
@@ -1,2 +1,2 @@
-export * from './lib/user.service';
 export * from './lib/create-user-dto';
+export * from './lib/user.service';
