+++ libs/menu-matriarch/settings/data-access/src/lib/user.service.ts
@@ -4,7 +4,7 @@ import { concatMap, first, tap } from 'rxjs/operators';
 
 import { AuthService } from '@atocha/firebase/data-access';
 import { User, UserPreferences } from '@atocha/menu-matriarch/shared/util';
-import { UserDataService } from './user-data.service';
+import { UserDataService } from './internal/user-data.service';
 
 @Injectable({
   providedIn: 'root',
