+++ libs/menu-matriarch/tags/data-access/src/index.ts
@@ -1,3 +1,3 @@
-export * from './lib/tag.service';
 export * from './lib/create-tag-dto';
+export * from './lib/tag.service';
 export * from './lib/mock-data';
