+++ libs/menu-matriarch/tags/data-access/src/lib/create-tag-dto.ts
@@ -1,4 +1,4 @@
-import { TagDto } from './tag-dto';
+import { TagDto } from './internal/tag-dto';
 
 export function createTagDto({
   id,
